Implement Datepicker Component & Cleanup Unused Styles
Removed unused styling defined in 'page.module.css' as it was affecting the overall aesthetics of the project. Introduced 'Calendar', 'Raw' and 'Day' components for better handling and representation of date functions, making use of 'date-fns' for date operations. Also made changes to the '.gitignore' file to exclude the '.idea' directory. Updated README.md to reflect the current project status. Lastly, added transparency effects for better visibility in the datepicker."

The changes primarily involved the implementation of a Datepicker component to enhance the functionality of the application. Unused styles were removed from the main style sheet to streamline the CSS and improve performance. Additionally, the '.gitignore' file was updated to ensure that the IDE configuration files are not tracked by Git. README was updated to be more direct and informative about the project. Transparency effects were added to datepicker for better UX.